+++ src/renderer.c
@@ -71,7 +71,7 @@ void render_sprite(u32 texture, i32 x, i32 y, i32 w, i32 h, float angle) {
   glBindVertexArray(0);
 }
 
-void render_texture_region(u32 texture, i32 x, i32 y, i32 w, i32 h, float angle, i32 x_offset, i32 y_offset, i32 x_range, i32 y_range, i32 texture_width, i32 texture_height) {
+void render_texture_region(u32 texture, float x, float y, float w, float h, float angle, i32 x_offset, i32 y_offset, i32 x_range, i32 y_range, i32 texture_width, i32 texture_height) {
   const u32 program = sprite_shader;
   glUseProgram(sprite_shader);
